\name{hongkong_area}
\docType{data}
\alias{hongkong_area}
\title{Area frequency table (meters squared) for Hong Kong}
\description{
A 2016 censored, univariate frequency table containing area (meters squared) for 87600 sub-divided units in Hong Kong. This table is formatted to look the same as a read in csv file [read.csv("filename.csv", row.names = NULL, header= FALSE, check.names=FALSE)].
}
\usage{hongkong_area}
\source{Thematic Household Survey Report - Report No. 60 - Housing conditions of sub-divided units in Hong Kong, 2016, page 23.}
\references{
Census and Statistics Department of Hong Kong Special Administrative Region . (2016). \emph{Thematic Household Survey Report - Report No. 60 - Housing conditions of sub-divided units in Hong Kong}.  Retrieved from: http://www.censtatd.gov.hk/hkstat/sub/sp100.jsp?productCode=C0000091
}
\keyword{datasets}


