\name{indonesia_contingency}
\docType{data}
\alias{indonesia_contingency}
\title{Contingency table of household size x area (meters squared) for Indonesia's rural Aceh Province}
\description{
A 2010 censored contingency table containing household size and area (meters squared) in Indonesia's rural Aceh Province. This table is formatted to look the same as a read in csv file [read.csv("filename.csv", row.names = NULL, header= TRUE, check.names=FALSE)].
}
\usage{indonesia_contingency}
\source{Population Census Data - Statistics Indonesia, 2010}
\references{
Population Census Data - Statistics Indonesia. (2010). \emph{Household by Floor Area of Dwelling Unit and Households Member Size}.  Retrieved from: http://sp2010.bps.go.id/index.php/site/tabel?wid=1100000000&tid=334&fi1=586&fi2=
}
\keyword{datasets}


