\name{iran_hhs}
\docType{data}
\alias{iran_hhs}
\title{Household size frequency table for East Azerbayejan (Azerbaijan), Iran}
\description{
A 2011 censored, univariate frequency table containing household size in East Azerbayejan (Azerbaijan), Iran.  This table is formatted to look the same as a read in csv file [read.csv("filename.csv", row.names = NULL, header= FALSE, check.names=FALSE)].
}
\usage{iran_hhs}
\source{Selected Findings of National Population and Housing Census, 2011, page 54.}
\references{
The Statistical Centre of Iran. (2011). \emph{Selected Findings of National Population and Housing Census}. Retrieved from: https://www.amar.org.ir/Portals/1/Iran/90.pdf
}
\keyword{datasets}




