\name{nepal_hhs}
\docType{data}
\alias{nepal_hhs}
\title{Household size frequency table for urban Nepal}
\description{
A 2011 censored, univariate frequency table containing household size in urban Nepal. This table is formatted to look the same as a read in csv file [read.csv("filename.csv", row.names = NULL, header= FALSE, check.names=FALSE)].
}
\usage{nepal_hhs}
\source{The Nepal Living Standards Survey, 2011, page 28.}
\references{
National Planning Commissions Secretariat, Government of Nepal. (2011). \emph{Nepal Living Standards Survey}. Retrieved from: http://siteresources.worldbank.org/INTLSMS/Resources/3358986-1181743055198/3877319-1329489437402/Statistical_Report_Vol1.pdf
}
\keyword{datasets}
